Fixed a number of analyze flagged warnings. Fixed mutable arrays being assigned where objectLists were expected. Fixed renderer background being assigned a UIColor rather than MBColor. Added Fade out timer on HUD background. Fixed in-app purchase loading of content where there is a user default receipt record but an invalid app receipt. Minor layout changes to the Welcome screen.

+++ FractalScape/model/LSDrawingRuleType.m
@@ -56,7 +56,7 @@
         
         NSInteger addRulesCount = [newDrawingRuleType loadRulesFromPListRulesArray: plistRulesArray];
         
-        NSLog(@"Added %ld rules.", addRulesCount);
+        NSLog(@"Added %ld rules.", (long)addRulesCount);
     }
     return newDrawingRuleType;
 }
